# Optimise merge where one side is a subset of the other