--- a/extras/generate_latest_map
+++ b/extras/generate_latest_map
@@ -96,14 +96,16 @@ function generate_latest_map() {
     loop:
         for i < len(m) {
             switch {
-            case j >= len(n) || m[i].key < n[j].key:
-                i++
+            case j >= len(n):
+                return m
             case m[i].key == n[j].key:
                 if m[i].Timestamp.Before(n[j].Timestamp) {
                     break loop
                 }
                 i++
                 j++
+            case m[i].key < n[j].key:
+                i++
             default:
                 break loop
             }
@@ -117,9 +119,9 @@ function generate_latest_map() {
 
         for i < len(m) {
             switch {
-            case j >= len(n) || m[i].key < n[j].key:
-                out = append(out, m[i])
-                i++
+            case j >= len(n):
+                out = append(out, m[i:]...)
+                return out
             case m[i].key == n[j].key:
                 if m[i].Timestamp.Before(n[j].Timestamp) {
                     out = append(out, n[j])
@@ -128,6 +130,9 @@ function generate_latest_map() {
                 }
                 i++
                 j++
+            case m[i].key < n[j].key:
+                out = append(out, m[i])
+                i++
             default:
                 out = append(out, n[j])
                 j++