This is beneficial because Scope very often merges multiple versions of the same node over time that have exactly the same labels, tags, etc., so we detect this and avoid copying.
This PR covers just three data structures - `StringSet`, `StringLatestMap` and `NodeControlDataLatestMap` - which benefit the most in profiles.

All three now use the same loop structure to merge. The loop is coded twice: once as far as we can get without changing the result, and then again after allocating space for a new return value.  If we got all the way to the end without changing anything, then return what we already have.

`StringSet` is mainly used in `Sets` where it benefits to avoid assigning to its map, so add an `unchanged` return.

Before (best of five):
```
BenchmarkReportMerge-2   	       3	 415559420 ns/op	110109765 B/op	  887819 allocs/op
BenchmarkReportUpgrade-2   	      20	  74962687 ns/op	 4983769 B/op	   22949 allocs/op
```

After:
```
BenchmarkReportMerge-2   	       5	 285569184 ns/op	63896876 B/op	  475350 allocs/op
BenchmarkReportUpgrade-2   	      20	  74528873 ns/op	 4856011 B/op	   18995 allocs/op
```