@rade 2018-07-04
I suggest cherry-picking de47a0eb0b93d1 (Less verbose string dump) onto master, since it's a bit of a distraction in this PR.